+++ lib/flutter_graph_widget.dart
@@ -16,18 +16,17 @@ class FlutterGraphWidget extends StatelessWidget {
   final Graph graph = Graph();
 
   FlutterGraphWidget({
-    Key? key,
+    super.key,
     required this.data,
     required this.convertor,
     required this.algorithm,
     Options? options,
-  })  : options = options ?? Options(),
-        super(key: key) {
+  }) : options = options ?? Options() {
     graph.options = options;
     this.options.graph = graph;
     graph.convertor = convertor;
     graph.algorithm = algorithm;
-    algorithm.setGlobalData(rootAlg: algorithm, graph: this.graph);
+    algorithm.setGlobalData(rootAlg: algorithm, graph: graph);
   }
 
   @override