+++ lib/model/graph.dart
@@ -129,13 +129,13 @@ class Graph<ID> {
 
   double get zoom => options?.scale.value ?? 1;
   ValueNotifier get scale => options!.scale;
-  get refreshData => options?.refreshData;
-  get mergeGraph => options?.mergeGraph;
+  void Function(dynamic data)? get refreshData => options?.refreshData;
+  void Function(dynamic graphData, {bool manual})? get mergeGraph =>
+      options?.mergeGraph;
 
-  ///
+  /// serialize the graph
   Map<String, dynamic> serialize() =>
     {
       "vertexes": { for (var v in vertexes) v.id as String: v.serialize() },
-      // "algorithm": algorithm?.serialize(),
     };
 }